fix: rename utils files to util
